Refactor logo imports, update authentication providers, and enhance feature descriptions
