+++ calendar-platform/app/components/landingPage/Navbar.tsx
@@ -1,6 +1,6 @@
 import Image from "next/image";
 import Link from "next/link";
-import Logo from "@/public/adnanlogo.svg";
+import Logo from "@/public/logo2.svg";
 
 
 import { ThemeToggle } from "../dashboard/ThemeToggle";
@@ -10,7 +10,7 @@ export function Navbar() {
     <div className="relative flex flex-col w-full py-5 mx-auto md:flex-row md:items-center md:justify-between">
       <div className="flex flex-row items-center justify-between text-sm lg:justify-start">
         <Link href="/" className="flex items-center gap-2">
-          <Image src={Logo}   className="size-10" alt="Logo" />
+          <Image src={Logo}   className="size-28" alt="Logo" />
 
         </Link>
         <div className="md:hidden">
