+++ calendar-platform/app/components/landingPage/oldhomepage.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import Image from "next/image";
 import Link from "next/link";
-import Logo from "@/public/adnanlogo.svg";
+import Logo from "@/public/logo.svg";
 
 export function OldHomePage() {
   return (
