+++ calendar-platform/package.json
@@ -14,7 +14,7 @@
     "@conform-to/zod": "^1.1.5",
     "@internationalized/date": "^3.5.5",
     "@nylas/react": "^1.1.3",
-    "@prisma/client": "^5.19.1",
+    "@prisma/client": "^5.22.0",
     "@radix-ui/react-dialog": "^1.1.1",
     "@radix-ui/react-dropdown-menu": "^2.1.1",
     "@radix-ui/react-icons": "^1.3.0",
@@ -30,7 +30,7 @@
     "date-fns": "^3.6.0",
     "iron-session": "^8.0.3",
     "lucide-react": "^0.440.0",
-    "next": "14.2.10",
+    "next": "^15.1.4",
     "next-auth": "^5.0.0-beta.21",
     "next-themes": "^0.3.0",
     "npm": "^10.8.3",
@@ -54,7 +54,7 @@
     "eslint": "^8",
     "eslint-config-next": "14.2.10",
     "postcss": "^8",
-    "prisma": "^5.19.1",
+    "prisma": "^5.22.0",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   }
