+++ calendar-platform/prisma/schema.prisma
@@ -1,17 +1,14 @@
-// This is your Prisma schema file,
-// learn more about it in the docs: https://pris.ly/d/prisma-schema
-
-// Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
-// Try Prisma Accelerate: https://pris.ly/cli/accelerate-init
-
 generator client {
   provider = "prisma-client-js"
+  previewFeatures = []
+  engineType = "binary"
 }
 
 datasource db {
   provider  = "postgresql"
   url       = env("DATABASE_URL")
   directUrl = env("DIRECT_URL")
+  relationMode = "prisma"
 }
 
 model User {
@@ -20,15 +17,15 @@ model User {
   email         String         @unique
   emailVerified DateTime?
   image         String?        @default("https://avatars.githubusercontent.com/u/10367109?v=4")
-  accounts      Account[]
-  sessions      Session[]
   grantId       String?
   grantEmail    String?
   username      String?        @unique
   createdAt     DateTime       @default(now())
   updatedAt     DateTime       @updatedAt
+  accounts      Account[]
   Availability  Availability[]
   EventType     EventType[]
+  sessions      Session[]
 }
 
 model Account {
@@ -43,11 +40,9 @@ model Account {
   scope             String?
   id_token          String?
   session_state     String?
-
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
-
-  user User @relation(fields: [userId], references: [id], onDelete: Cascade)
+  createdAt         DateTime @default(now())
+  updatedAt         DateTime @updatedAt
+  user              User     @relation(fields: [userId], references: [id], onDelete: Cascade)
 
   @@id([provider, providerAccountId])
 }
@@ -56,10 +51,9 @@ model Session {
   sessionToken String   @unique
   userId       String
   expires      DateTime
+  createdAt    DateTime @default(now())
+  updatedAt    DateTime @updatedAt
   user         User     @relation(fields: [userId], references: [id], onDelete: Cascade)
-
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
 }
 
 model VerificationToken {
@@ -71,18 +65,28 @@ model VerificationToken {
 }
 
 model Availability {
-  id String @id @default(uuid())
-
-  day      Day
-  fromTime String // e.g., '08:00'
-  tillTime String // e.g., '18:00'
-  isActive Boolean @default(true) // Toggle availability for the day
-
-  User   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
-  userId String
-
+  id        String   @id @default(uuid())
+  day       Day
+  fromTime  String
+  tillTime  String
+  isActive  Boolean  @default(true)
+  userId    String
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
+  User      User     @relation(fields: [userId], references: [id], onDelete: Cascade)
+}
+
+model EventType {
+  id                String   @id @default(uuid())
+  title             String
+  duration          Int
+  url               String
+  description       String
+  active            Boolean  @default(true)
+  videoCallSoftware String   @default("Google Meet")
+  userId            String
+  createdAt         DateTime @default(now())
+  user              User     @relation(fields: [userId], references: [id], onDelete: Cascade)
 }
 
 enum Day {
@@ -93,19 +97,4 @@ enum Day {
   Friday
   Saturday
   Sunday
-}
-
-model EventType {
-  id                String  @id @default(uuid())
-  title             String
-  duration          Int
-  url               String
-  description       String
-  active            Boolean @default(true)
-  videoCallSoftware String  @default("Google Meet")
-
-  user   User   @relation(fields: [userId], references: [id], onDelete: Cascade)
-  userId String
-
-  createdAt DateTime @default(now())
 }
