+++ calendar-platform/tailwind.config.ts
@@ -1,5 +1,4 @@
 import type { Config } from "tailwindcss";
-import { withUt } from "uploadthing/tw";
 
 const config: Config = {
   darkMode: ["class"],
@@ -59,6 +58,7 @@ const config: Config = {
       },
     },
   },
-  plugins: [require("tailwindcss-animate")],
-};
-export default withUt(config);
+  plugins: [import("tailwindcss-animate")],
+} satisfies Config;
+
+export default config;
